// model/changeSite200ResponseAllOfDataInner.ts
/**
 * ManageSpace Search API
 * ManageSpace Search API Documentation
 *
 * The version of the OpenAPI document: 0.0.12-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ChangeSite200ResponseAllOfDataInner {
    'siteName'?: string | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "siteName",
            "baseName": "siteName",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ChangeSite200ResponseAllOfDataInner.attributeTypeMap;
    }
}