// model/contactAdd200ResponseAllOfDataInner.ts
/**
 * ManageSpace Search API
 * ManageSpace Search API Documentation
 *
 * The version of the OpenAPI document: 0.0.12-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ContactAdd200ResponseAllOfDataInner {
    'accountName'?: string | null;
    'companyName'?: string | null;
    'firstName'?: string | null;
    'lastName'?: string | null;
    /**
    * Phone number in E.164 phone number format (assumed +1 if no international code provided)
    */
    'phoneNumber'?: string | null;
    /**
    * Mobile number in E.164 phone number format (assumed +1 if no international code provided)
    */
    'mobileNumber'?: string | null;
    'email'?: string | null;
    'addressLine1'?: string | null;
    'addressLine2'?: string | null;
    'city'?: string | null;
    'state'?: string | null;
    'country'?: string | null;
    /**
    * A correctly formatted zip/postal code to match the country
    */
    'zipCode'?: string | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "accountName",
            "baseName": "accountName",
            "type": "string"
        },
        {
            "name": "companyName",
            "baseName": "companyName",
            "type": "string"
        },
        {
            "name": "firstName",
            "baseName": "firstName",
            "type": "string"
        },
        {
            "name": "lastName",
            "baseName": "lastName",
            "type": "string"
        },
        {
            "name": "phoneNumber",
            "baseName": "phoneNumber",
            "type": "string"
        },
        {
            "name": "mobileNumber",
            "baseName": "mobileNumber",
            "type": "string"
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "addressLine1",
            "baseName": "addressLine1",
            "type": "string"
        },
        {
            "name": "addressLine2",
            "baseName": "addressLine2",
            "type": "string"
        },
        {
            "name": "city",
            "baseName": "city",
            "type": "string"
        },
        {
            "name": "state",
            "baseName": "state",
            "type": "string"
        },
        {
            "name": "country",
            "baseName": "country",
            "type": "string"
        },
        {
            "name": "zipCode",
            "baseName": "zipCode",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ContactAdd200ResponseAllOfDataInner.attributeTypeMap;
    }
}