// model/feeAdd200ResponseAllOfDataInner.ts
/**
 * ManageSpace Search API
 * ManageSpace Search API Documentation
 *
 * The version of the OpenAPI document: 0.0.12-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class FeeAdd200ResponseAllOfDataInner {
    'feeName'?: string | null;
    'rateType'?: FeeAdd200ResponseAllOfDataInner.RateTypeEnum;
    /**
    * For fixedRate, an amount in the smallest denomination (e.g. cents/pence)
    */
    'amount'?: number | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "feeName",
            "baseName": "feeName",
            "type": "string"
        },
        {
            "name": "rateType",
            "baseName": "rateType",
            "type": "FeeAdd200ResponseAllOfDataInner.RateTypeEnum"
        },
        {
            "name": "amount",
            "baseName": "amount",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return FeeAdd200ResponseAllOfDataInner.attributeTypeMap;
    }
}

export namespace FeeAdd200ResponseAllOfDataInner {
    export enum RateTypeEnum {
        Percentage = <any> 'percentage',
        FixedRate = <any> 'fixedRate'
    }
}